chore: rework async toggling to improve usability
I wasn't a fan of where the `AsyncToggleable` class ended up. Felt like combinging toggle state management and resource ownership was a mistake, since the indirection made the code hard to reason about. This change largely untangles that and completely re-works the tests to hopefully be more readable; facilitate by the addition of a _testing_ `async` helper which effectively wraps/spys on the invocation of the async methods for a toggle.

I'll be honest, I'm still not totally happy with this code, but am going to prioritize switching gears to some more important problems. Down the road, we should consider reworking this or even leaning on something like RxJS which can facilitate doing async work triggered by a sync event (why this exists in the first place). Something like a `switchMap` could be a good fit.

wip: simplify eventing/toggling/running async tasks from sync listeners